+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-portal
 0.57.10 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- Added preparations for themakaart maatregelen.
 
 
 0.57.9 (2012-05-30)